Started development. Added first structure and tabview quick test
