--- DogFinder/DogFinder/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  DogFinder
-//
-//  Created by José Marques on 14/01/2023.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
